IA/Reconnaissance vocale/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-SocketIO==5.5.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
numpy==2.2.2
openai==1.58.1
paho-mqtt==2.1.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.0
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
